refactor(scripts): #555 PR-B rename for clarity + add README index
- pre-deploy-check.sh → pre_deploy_check.sh (kebab → snake)
- auto_deploy.sh → autopull_receiver.sh (was misleading: receiver, not push)
- deploy.sh → deploy_remote.sh (clarity)
- deploy_mini.sh → deploy_to_mini.sh (clarity)
- db_maintenance.py → maintenance.py (drop redundant prefix in db/)
- migrate_db.py → migrate.py
- restore_db.sh → restore.sh
- scripts/README.md (NEW): comprehensive index with sub-dir tables

Updates Makefile / docs / launchd plist / scheduler / tests references.
Tests: 4462/4462 green.
