+++ scripts/deploy/pre_deploy_check.sh
